+++ src/devguide/fulcro_devguide/putting_together/soln_ex_1.cljs
@@ -4,11 +4,11 @@
   (:require [om.next :as om :refer-macros [defui]]
             [om.dom :as dom]
             [devcards.core :as dc :refer-macros [defcard defcard-doc]]
-            [fulcro.client.core :as uc]
+            [fulcro.client.core :as fc]
             [fulcro.client.data-fetch :as df]))
 
 (defui ^:once TodoItem
-  static uc/InitialAppState
+  static fc/InitialAppState
   (initial-state [this {:keys [id label complete]}] {:item/id id :item/label label :item/done complete})
   static om/IQuery
   (query [this] [:item/id :item/label :item/done])
@@ -26,11 +26,11 @@
 (def ui-item (om/factory TodoItem))
 
 (defui ^:once ItemList
-  static uc/InitialAppState
-  (initial-state [this params] {:list/title "My List" :list/items [(uc/initial-state TodoItem {:id 1 :label "A" :complete false})
-                                                                   (uc/initial-state TodoItem {:id 2 :label "B" :complete false})
-                                                                   (uc/initial-state TodoItem {:id 3 :label "C" :complete true})
-                                                                   (uc/initial-state TodoItem {:id 4 :label "D" :complete false})]})
+  static fc/InitialAppState
+  (initial-state [this params] {:list/title "My List" :list/items [(fc/initial-state TodoItem {:id 1 :label "A" :complete false})
+                                                                   (fc/initial-state TodoItem {:id 2 :label "B" :complete false})
+                                                                   (fc/initial-state TodoItem {:id 3 :label "C" :complete true})
+                                                                   (fc/initial-state TodoItem {:id 4 :label "D" :complete false})]})
   static om/IQuery
   (query [this] [:list/title {:list/items (om/get-query TodoItem)}])
   static om/Ident
@@ -46,8 +46,8 @@
 (def ui-item-list (om/factory ItemList))
 
 (defui ^:once TodoList
-  static uc/InitialAppState
-  (initial-state [this params] {:ui/react-key "A" :item-list (uc/initial-state ItemList nil)})
+  static fc/InitialAppState
+  (initial-state [this params] {:ui/react-key "A" :item-list (fc/initial-state ItemList nil)})
   static om/IQuery
   (query [this] [:ui/react-key {:item-list (om/get-query ItemList)}])
   Object