+++ src/devguide/fulcro_devguide/putting_together/soln_ex_3.cljs
@@ -4,7 +4,7 @@
   (:require [om.next :as om :refer-macros [defui]]
             [om.dom :as dom]
             [devcards.core :as dc :refer-macros [defcard defcard-doc]]
-            [fulcro.client.core :as uc]
+            [fulcro.client.core :as fc]
             [fulcro.client.data-fetch :as df]
             [fulcro.client.mutations :as m]))
 
@@ -29,10 +29,10 @@
              (let [new-ident [:items/by-id id]
                    new-item {:item/id id :item/label label :item/done false}]
                (swap! state assoc-in [:items/by-id id] new-item)
-               (uc/integrate-ident! state new-ident :append [:lists/by-title list :list/items])))})
+               (fc/integrate-ident! state new-ident :append [:lists/by-title list :list/items])))})
 
 (defui ^:once TodoItem
-  static uc/InitialAppState
+  static fc/InitialAppState
   (initial-state [this {:keys [id label complete]}] {:item/id id :item/label label :item/done complete})
   static om/IQuery
   (query [this] [:item/id :item/label :item/done])
@@ -51,7 +51,7 @@
 (def ui-item (om/factory TodoItem))
 
 (defui ^:once ItemList
-  static uc/InitialAppState
+  static fc/InitialAppState
   (initial-state [this params] {:list/title "My List" :list/items []})
   static om/IQuery
   (query [this] [:ui/new-item-text :list/title {:list/items (om/get-query TodoItem)}])
@@ -77,8 +77,8 @@
 (def ui-item-list (om/factory ItemList))
 
 (defui ^:once TodoList
-  static uc/InitialAppState
-  (initial-state [this params] {:ui/react-key "A" :item-list (uc/initial-state ItemList nil)})
+  static fc/InitialAppState
+  (initial-state [this params] {:ui/react-key "A" :item-list (fc/initial-state ItemList nil)})
   static om/IQuery
   (query [this] [:ui/react-key {:item-list (om/get-query ItemList)}])
   Object