+++ CHANGELOG.md
@@ -173,8 +173,8 @@ upload support)
 - Added `(start [this app])` to the `UntangledNetwork` protocol.
 - Log-app-state now requires the atom containing a mounted untangled client, define it in the user namespace like so:
 ```
-(defonce app (atom (uc/new-untangled-client ... )))
-(swap! app uc/mount RootComponent "app-div")
+(defonce app (atom (fc/new-untangled-client ... )))
+(swap! app fc/mount RootComponent "app-div")
 (def log-app-state (partial util/log-app-state app))
 ```
 - `global-error-callback` now expectes an arity 2 function. First param is the status and the second is the response.